Fixed some issues with Texture loading, added zoom factor in camera

+++ src/OGL_Camera.cpp
@@ -10,7 +10,7 @@ OGL_Camera* OGL_CreateCamera(glm::vec3 pos, glm::vec3 up, float yaw, float pitch
     newCam->front = glm::vec3(0.0f, 0.0f, -1.0f);
 
     newCam->fov = 45.0f;
-    newCam->aspect = SDL2_WinWidth / SDL2_WinHeight;
+    newCam->aspect = static_cast<float>(SDL2_WinWidth) / static_cast<float>(SDL2_WinHeight); // Better to update than to cache it
     newCam->nearPlane = 0.1f;
     newCam->farPlane = 100.0f;
